Add foreground element manipulation methods
Added three methods to manipulate foreground elements: specifically a method for adding, removing, and moving foreground elements in the VNScreen. Also added a script to the foreground_element scene, into which I added functionality for linear movement (in the _process method) and a method for setting linear movement. Further details are in code comments.